UpTraderTask/requirements.txt
Django==2.1.4
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
djangorestframework-recursive==0.1.2
pytz==2018.7
